refactor(discord): extract cancel and prompt replacement arms (S9)
