+++ scripts/audit_maintainability_giant_baseline.toml
@@ -227,7 +227,7 @@
 # warning block moved verbatim to the early_tui_completion.rs sibling module
 # (behavior-preserving decompose; the two outputs are threaded back by return).
 "src/services/discord/turn_bridge/mod.rs" = 1232 # 2026-07-08 #4230 S6 review fix (E0425): 1230 -> 1232 (+2 prod LoC), reviewable admission — SPINNER + LIVE_LONG_RUN_HEARTBEAT_INTERVAL hoisted to module scope for the extracted stream_loop.rs; +2 is the placement comment, const values unchanged. # 2026-07-08 #4230 S6: 2492 -> 1230 (-1262 prod LoC), ratchet down after extracting the main stream loop shell + remaining event arms to turn_bridge/stream_loop.rs (behavior-preserving decompose; external return/break/continue mappings=0/0/0; internal labelled break=6, inner break=3, continue=1). # 2026-07-08 #4230 S5: 3831 -> 2492 (-1339), ratchet down after verbatim-move of the terminal outcome delivery block to turn_bridge/terminal_outcome_delivery.rs (behavior-preserving decompose; external return/break/continue mappings=0/0/0, internal-loop continue=1) # 2026-07-08 #4230 S4: 4326 -> 3831 (-495), ratchet down after verbatim-move of the post-loop owner classification + finalizer block to turn_bridge/post_loop_finalize.rs (behavior-preserving decompose) # 2026-07-08 #4230 S3: 5019 -> 4326 (-693), ratchet down after verbatim-move of the completion postlude + inflight epilogue (incl. #4185 restart-cancel block intact) to turn_bridge/completion_postlude.rs (behavior-preserving decompose) # 2026-07-08 #4230 S2: 5692 -> 5019 (-673), ratchet down after verbatim-move of runtime handoff arms + handle_watcher_runtime_handoff to turn_bridge/runtime_handoff_loop.rs (behavior-preserving decompose) # 2026-07-08 #4230 S1: ceiling 6250 -> 5692 (actual prod 6193 -> 5692, -501; slack removed), ratchet down after verbatim-move of CompletionGuard/InflightCleanupGuard to turn_bridge/guards.rs and the post-drain stream/status tick to turn_bridge/stream_tick.rs (behavior-preserving decompose) # 2026-07-03 #3998 S1-f2: 6284 -> 6250 (-34 prod LoC), ratchet down after retiring the A5/A6b rollout OR-in and routing site-5 from structural A5 inputs only. # 2026-07-02 #3805 P2 PR-D: 6241 -> 6284 (+43 prod LoC), reviewable admission — rollover re-anchor of the two-message status panel (send new panel below the new tail answer + retire old + in-memory epoch bump persisted by the loop's save_inflight_state flock); all logic lives in the non-giant turn_bridge/two_message_panel.rs, mod.rs carries ONLY a per-interval rolled-over local + one gated re-anchor call after the rollover loop. Gated on default-OFF two_message_panel_enabled → OFF byte-identical; ordering / finalize / footer untouched. # 2026-07-02 #3805 P2 PR-C: 6237 -> 6241 (+4 LoC), reviewable admission — re-export `two_message_status_edit_generation_is_stale` at `pub(in crate::services::discord)` so the tmux WATCHER completion guard reuses the SAME sink generation-staleness predicate (parity). Wiring-only: one `pub(in ...) use two_message_panel::...;` line + its comment; no logic change, OFF path byte-identical. # 2026-07-02 #3805 P2 PR-B: 6223 -> 6237 (+14 prod LoC), reviewable admission — two-message sink creation order (answer-first, panel below). The two-message create + generation staleness guard live entirely in the new non-giant turn_bridge/two_message_panel.rs (120 prod LoC); mod.rs carries only the thin wiring — a `mod two_message_panel;` decl, a per-turn `status_panel_generation` local seeded from the inflight snapshot, the `two_message_panel_enabled` flag + `&mut status_panel_generation` args on the existing `maybe_create_bridge_separate_status_panel_response` call, and the `status_panel_generation` arg on the terminal completion call. Gated on `two_message_panel_enabled` (default OFF) → OFF path byte-identical; ordering / finalize / rollover / discord_io untouched. # 2026-07-02 #3983 item4: 6222 -> 6223 (+1 prod LoC), reviewable admission — one-shot top session banner; only the `&provider` argument is threaded into the existing sink `refresh_session_panel_line_from_lifecycle` call, the emit + de-dup live in the new non-giant session_banner.rs + panel_lifecycle.rs. # #3607 terminal-UI obligation hook # #3813 Phase 1b: 6201 -> 6209 (+8), reviewable admission — first-output fast-lane status-edit gate. The streaming status-edit gate held the first assistant answer up to status_interval (default 5s); a single-shot fast lane now relays the FIRST non-empty assistant text immediately, then reverts to the normal throttle (turn-level cost is at most +1 edit). The pure decision lives in streaming_edit_text.rs (bridge_streaming_edit_gate_open, unit-tested); mod.rs carries only the thin call-site: a first_answer_relayed flag decl, the gate subcondition swap, and the flag set on a successful non-empty edit. flicker suppression / discord_io min_gap / ordering / finalize (!done guard, rollover, ownership counters) untouched. # #3813 Phase 2 + Bridge-spans: 6209 -> 6235 (+26), reviewable admission — same +26 as the raw ratchet since every added line is production body. Phase 2 (§3 status-panel low-pri): defers the v2 status-panel / footer edit off the shared per-channel rate lane while the opening answer is un-relayed so the #4006 fast lane wins it; pure predicate status_panel_edit_defer_for_first_answer in single_message_footer.rs (unit-tested), mod.rs adds first_answer_text_pending compute + one gate subcondition on each of the separate-panel / footer blocks. #3477 guarded (deferral needs a genuinely-pending first-answer body → tool-only / watcher-owned turns never suppressed; status_panel_dirty preserved → renders next interval). Bridge-spans (AC#1 tail): observation-only bridge latency spans reusing the turn_start anchor (no signature / Context change); struct + emit in the new bridge_latency_spans.rs, mod.rs carries only the decl + idempotent first_output/first_relay marks + one loop-exit emit. discord_io min_gap / ordering / finalize / flicker untouched; no new await or lock.
-"src/services/discord/turn_bridge/terminal_outcome_delivery.rs" = 1669 # 2026-07-08 #4230 S5: new high-coupling terminal outcome delivery child registered at actual production LoC after the approved contiguous verbatim extraction from turn_bridge/mod.rs (external return/break/continue mappings=0/0/0; internal-loop continue=1).
+"src/services/discord/turn_bridge/terminal_outcome_delivery.rs" = 1370 # 2026-07-10 #4230 S9: 1669 -> 1370 (-299 prod LoC), ratchet down after extracting the cancelled/prompt-too-long terminal replacement arms to terminal_outcome_delivery/cancel_prompt_replace.rs with both #3041 delivery-lease acquire/commit pairs intact (external labelled-break/break/continue/return mappings=0/0/0/0). # 2026-07-08 #4230 S5: new high-coupling terminal outcome delivery child registered at actual production LoC after the approved contiguous verbatim extraction from turn_bridge/mod.rs (external return/break/continue mappings=0/0/0; internal-loop continue=1).
 # #3038 turn_bridge S1 child modules, frozen at landed production LoC.
 "src/services/discord/turn_bridge/headless_delivery.rs" = 415
 "src/services/discord/turn_bridge/status_panel.rs" = 437
